script/requirements.txt
numpy==2.0.0
pillow==10.4.0
